Update and use useIsMounted hook
