--- src/frontend/hooks/useIsMounted.js
@@ -1,19 +1,0 @@
-// @flow
-import { useRef, useEffect, useCallback } from "react";
-
-/**
- * const isMounted = useIsMounted()
- * console.log(isMounted())
- */
-export default function useIsMounted(): () => boolean {
-  const ref = useRef(false);
-
-  useEffect(() => {
-    ref.current = true;
-    return () => {
-      ref.current = false;
-    };
-  }, []);
-
-  return useCallback(() => ref.current, []);
-}
