+++ src/frontend/hooks/useIsMounted.ts
@@ -1,0 +1,14 @@
+import { useCallback, useEffect, useRef } from "react";
+
+export default function useIsMounted() {
+  const ref = useRef(false);
+
+  useEffect(() => {
+    ref.current = true;
+    return () => {
+      ref.current = false;
+    };
+  }, []);
+
+  return useCallback(() => ref.current, []);
+}
