+++ src/frontend/screens/ProjectInviteModal.tsx
@@ -10,6 +10,7 @@ import { NavigationStackScreenComponent } from "react-navigation-stack";
 import { defineMessages, useIntl } from "react-intl";
 import MaterialIcon from "react-native-vector-icons/MaterialIcons";
 
+import useIsMounted from "../hooks/useIsMounted";
 import { DARK_BLUE, MAGENTA, MAPEO_BLUE, WHITE } from "../lib/styles";
 import {
   MODAL_NAVIGATION_OPTIONS,
@@ -102,9 +103,8 @@ export const ProjectInviteModal: NavigationStackScreenComponent<{
   invite?: string;
 }> = ({ navigation }) => {
   const { formatMessage: t } = useIntl();
+  const isMounted = useIsMounted();
   const { sheetRef, closeSheet } = useBottomSheetModal({ openOnMount: true });
-
-  const mountedRef = React.useRef(true);
 
   const inviteKey = navigation.getParam("invite");
 
@@ -116,21 +116,24 @@ export const ProjectInviteModal: NavigationStackScreenComponent<{
       : { type: "error", info: { error: new Error(t(m.inviteErrorMessage)) } }
   );
 
-  const fetchInviteDetails = React.useCallback(async (inviteKey: string) => {
-    try {
-      setStatus({ type: "loading" });
+  const fetchInviteDetails = React.useCallback(
+    async (inviteKey: string) => {
+      try {
+        setStatus({ type: "loading" });
 
-      const inviteDetails = await getProjectInviteDetails(inviteKey);
+        const inviteDetails = await getProjectInviteDetails(inviteKey);
 
-      if (mountedRef.current) {
-        setStatus({ type: "success", info: { inviteDetails } });
+        if (isMounted()) {
+          setStatus({ type: "success", info: { inviteDetails } });
+        }
+      } catch (err) {
+        if (isMounted() && err instanceof Error) {
+          setStatus({ type: "error", info: { error: err } });
+        }
       }
-    } catch (err) {
-      if (mountedRef.current && err instanceof Error) {
-        setStatus({ type: "error", info: { error: err } });
-      }
-    }
-  }, []);
+    },
+    [isMounted]
+  );
 
   const acceptInvite = () => {
     navigation.navigate("Sync");
@@ -141,12 +144,6 @@ export const ProjectInviteModal: NavigationStackScreenComponent<{
       fetchInviteDetails(inviteKey);
     }
   }, [fetchInviteDetails, inviteKey]);
-
-  React.useEffect(() => {
-    return () => {
-      mountedRef.current = false;
-    };
-  }, []);
 
   return (
     <BottomSheetModal ref={sheetRef} onDismiss={navigation.goBack}>
